+++ dbflute_maihamadb/lrevworld-test.sh
@@ -3,7 +3,12 @@
 cd `dirname $0`
 . _project.sh
 
-rm -Rf ./output/doc/data
+FIRST_ARG=$1
+
+if [ "$FIRST_ARG" = "clean" ];then
+  echo "...Cleaning existing resources"
+  rm -Rf ./output/doc/data
+fi
 
 export answer=y
