+++ dbflute_maihamadb/syncworld-test.sh
@@ -3,13 +3,19 @@
 cd `dirname $0`
 . _project.sh
 
+FIRST_ARG=$1
+
+if [ "$FIRST_ARG" = "clean" ];then
+  echo "...Cleaning existing resources"
+  rm ./output/doc/syncworld-sync-check-result.html
+fi
+
 export answer=y
 
 export DBFLUTE_ENVIRONMENT_TYPE=diffready
 
 . replace-schema.sh
 
-rm ./output/doc/syncworld-sync-check-result.html
 rm ./dfprop/syncworld/syncworlddb.mv.db
 cp ./dfprop/diffworld/diffworlddb.mv.db ./dfprop/syncworld/syncworlddb.mv.db
